+++ Minimal/Code/Enums/Entities.py
@@ -8,7 +8,7 @@ class Entity:
 class Entities(_Enum):
 	@staticmethod
 	def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[_Any]) -> Entity:
-		return Entity()
+		...
 
 	Apple: Entity
 	# Docstring: Apple (Entities)