+++ Minimal/Code/Enums/Hats.py
@@ -8,7 +8,7 @@ class Hat:
 class Hats(_Enum):
 	@staticmethod
 	def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[_Any]) -> Hat:
-		return Hat()
+		...
 
 	Brown_Hat: Hat
 	# Docstring: Brown_Hat