+++ Minimal/Code/Enums/Items.py
@@ -8,7 +8,7 @@ class Item:
 class Items(_Enum):
     @staticmethod
     def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[_Any]) -> Item:
-        return Item()
+        ...
 
     Bone: Item
     # Docstring: Bone (Items)